restaurant_api/menu/serializers.py
from rest_framework import serializers
from restaurant_api.menu.models import Menu


class MenuSerializer(serializers.ModelSerializer):
    class Meta:
        model = Menu
        fields = '__all__'
