restaurant_api/requirements.txt
django >= 5.0.5
djangorestframework >= 3.15.1
psycopg2-binary >= 2.9.9
psycopg2 >= 2.9.9
pytest >= 8.2.0